# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Civic Co-Pilot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AdminUser {
  username: string;
  role: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  user: AdminUser;
}

export interface Complaint {
  id: number;
  title: string;
  description: string;
  category: string;
  status: string;
  priority: string;
  citizenName: string;
  /** @nullable */
  citizenEmail?: string | null;
  /** @nullable */
  citizenPhone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ComplaintInput {
  title: string;
  description: string;
  category: string;
  priority: string;
  citizenName: string;
  citizenEmail?: string;
  citizenPhone?: string;
  location?: string;
}

export interface ComplaintUpdate {
  status?: string;
  priority?: string;
  adminNotes?: string;
  title?: string;
  description?: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface ComplaintStats {
  total: number;
  pending: number;
  inReview: number;
  resolved: number;
  rejected: number;
  highPriority: number;
  recentCount: number;
  byCategory: CategoryCount[];
}

export interface Scheme {
  id: number;
  title: string;
  description: string;
  category: string;
  /** @nullable */
  eligibility?: string | null;
  link: string;
  /** @nullable */
  ministry?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}

export interface SchemeInput {
  title: string;
  description: string;
  category: string;
  eligibility?: string;
  link: string;
  ministry?: string;
  state?: string;
  tags?: string;
}

export interface SchemeUpdate {
  title?: string;
  description?: string;
  category?: string;
  eligibility?: string;
  link?: string;
  ministry?: string;
  state?: string;
  tags?: string;
}

export type ListComplaintsParams = {
status?: string;
category?: string;
priority?: string;
search?: string;
};

export type ListSchemesParams = {
category?: string;
search?: string;
};