# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Civic Co-Pilot API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: complaints
    description: Complaint management
  - name: schemes
    description: Government schemes

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List all complaints
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Submit a new complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

  /complaints/stats:
    get:
      operationId: getComplaintStats
      tags: [complaints]
      summary: Get complaint statistics for dashboard
      responses:
        "200":
          description: Complaint statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintStats"

  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get a complaint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint status and notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: Updated complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    delete:
      operationId: deleteComplaint
      tags: [complaints]
      summary: Delete a complaint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /schemes:
    get:
      operationId: listSchemes
      tags: [schemes]
      summary: List government schemes
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of schemes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scheme"
    post:
      operationId: createScheme
      tags: [schemes]
      summary: Create a government scheme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchemeInput"
      responses:
        "201":
          description: Scheme created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scheme"

  /schemes/{id}:
    get:
      operationId: getScheme
      tags: [schemes]
      summary: Get a scheme by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scheme detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scheme"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateScheme
      tags: [schemes]
      summary: Update a scheme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchemeUpdate"
      responses:
        "200":
          description: Updated scheme
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scheme"
    delete:
      operationId: deleteScheme
      tags: [schemes]
      summary: Delete a scheme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    AdminUser:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
      required:
        - username
        - role

    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/AdminUser"
      required:
        - success
        - user

    Complaint:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
        priority:
          type: string
        citizenName:
          type: string
        citizenEmail:
          type: ["string", "null"]
        citizenPhone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - status
        - priority
        - citizenName
        - createdAt
        - updatedAt

    ComplaintInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        priority:
          type: string
        citizenName:
          type: string
        citizenEmail:
          type: string
        citizenPhone:
          type: string
        location:
          type: string
      required:
        - title
        - description
        - category
        - priority
        - citizenName

    ComplaintUpdate:
      type: object
      properties:
        status:
          type: string
        priority:
          type: string
        adminNotes:
          type: string
        title:
          type: string
        description:
          type: string

    ComplaintStats:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        inReview:
          type: integer
        resolved:
          type: integer
        rejected:
          type: integer
        highPriority:
          type: integer
        recentCount:
          type: integer
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
      required:
        - total
        - pending
        - inReview
        - resolved
        - rejected
        - highPriority
        - recentCount
        - byCategory

    CategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required:
        - category
        - count

    Scheme:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        eligibility:
          type: ["string", "null"]
        link:
          type: string
        ministry:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - category
        - link
        - createdAt

    SchemeInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        eligibility:
          type: string
        link:
          type: string
        ministry:
          type: string
        state:
          type: string
        tags:
          type: string
      required:
        - title
        - description
        - category
        - link

    SchemeUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        eligibility:
          type: string
        link:
          type: string
        ministry:
          type: string
        state:
          type: string
        tags:
          type: string