# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Civic Co-Pilot API
 * OpenAPI spec version: 0.1.0
 */

export interface Complaint {
  id: number;
  title: string;
  description: string;
  category: string;
  status: string;
  priority: string;
  citizenName: string;
  /** @nullable */
  citizenEmail?: string | null;
  /** @nullable */
  citizenPhone?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
  updatedAt: string;
}