# lib/api-zod/src/generated/types/complaintInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Civic Co-Pilot API
 * OpenAPI spec version: 0.1.0
 */

export interface ComplaintInput {
  title: string;
  description: string;
  category: string;
  priority: string;
  citizenName: string;
  citizenEmail?: string;
  citizenPhone?: string;
  location?: string;
}