# lib/api-zod/src/generated/types/complaintStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Civic Co-Pilot API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';

export interface ComplaintStats {
  total: number;
  pending: number;
  inReview: number;
  resolved: number;
  rejected: number;
  highPriority: number;
  recentCount: number;
  byCategory: CategoryCount[];
}