# lib/api-zod/src/generated/types/complaintUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Civic Co-Pilot API
 * OpenAPI spec version: 0.1.0
 */

export interface ComplaintUpdate {
  status?: string;
  priority?: string;
  adminNotes?: string;
  title?: string;
  description?: string;
}