# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Civic Co-Pilot API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from './adminUser';

export interface LoginResponse {
  success: boolean;
  user: AdminUser;
}