# lib/api-zod/src/generated/types/scheme.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Civic Co-Pilot API
 * OpenAPI spec version: 0.1.0
 */

export interface Scheme {
  id: number;
  title: string;
  description: string;
  category: string;
  /** @nullable */
  eligibility?: string | null;
  link: string;
  /** @nullable */
  ministry?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
}