# lib/api-zod/src/generated/types/schemeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Civic Co-Pilot API
 * OpenAPI spec version: 0.1.0
 */

export interface SchemeInput {
  title: string;
  description: string;
  category: string;
  eligibility?: string;
  link: string;
  ministry?: string;
  state?: string;
  tags?: string;
}